Move type-level operations from Utils.Links to API.TypeLevel
